requirements.txt
pandas
snowflake-connector-python
streamlit
pyarrow<8.1.0,>=8.0.0
plotly
geopy